requirements.txt
Django==5.0.6
django-storages[google]==1.14.3
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
djangorestframework==3.15.1
pydantic==2.7.1
django-cors-headers==4.3.1
django-structlog==8.0.0
structlog==24.1.0
rich==13.7.1
google-cloud-secret-manager==2.20.0